#5688 Tell CSV search callers whether their results are complete
